qt: relax made current assertions in `CompositorX11UISurface`
Rather than having an assertion regarding the context became current,
we can simply early return in these functions. If the context could
not be made current, it is likely that we are in some sort of
transitional phase and `resizeFbo()` would be called later when the
state is stable either directly or indirectly.

`render()`, on the other is periodically called anyway. It should
be fine to simply return early instead of triggering assertion
failure. When the context can be made current, it can do its job.