qt: do not hide the interface window when adjusting `Qt::FramelessWindowHint`
- Since hiding the window destroys the wl_surface (before Qt 6.9), this causes
crash on Wayland (adjustments after video window setup is complete).
- With `CompositorX11`, adjusting a flag (such as always on top) and then
(immediately after) toggling the visibility of a window breaks the window. This
is a problem when, for example always on top and client side decorations change
together.

Although it is stated that "some window managers don't like to change frame
window hint on visible window", I have not observed an adverse behavior on X11
(Qt 6.2 and Qt 6.8), and Wayland (Qt 6.8) with KWin X11/Wayland 6.3.1. If this
is a problem of a certain window manager, the workaround should be only done
for that window manager and not in all cases.

At the same time, Qt does not seem to state that adjusting `Qt::FramelessWindowHint`
needs to be done when the window is hidden. So I believe that this should be
safe to proceed.