Revert "qml: use QSGTextureView (TextureProviderItem) in PartialEffect instead of layering"
This reverts commit 60e95ad5c2ab53dec17112fc9caab6bf52a90db8.

Unfortunately there is not much to do here until `MultiEffect` starts supporting sub-
textures (textures that are independent/not in atlas, but report position different
than (0.0, 0.0) or size different than (1.0, 1.0)).

Although this is only relevant for `MultiEffect` (because other type of effects can
set `supportsAtlasTextures` or use `qt_SubRect_x`), I still reverted it to not impair
maintenance. I also do not want to bring back manually setting `supportsAtlasTextures`
to `true`.

We should still keep the texture view and provider, because they are useful when sub-
texturing is not used (such as player view), and when the consumer supports sub-textures.