direct3d11: separate display from drawable display
Separating the display used in VLC and the libvlc application-side
display ensures we're not trying to use the latter in the VLC scenarios
and still ensures we check the kind of windowing that is being provided
to the display.

Fix a regression with e33e383fde533c17032614bd0cbf5be1d364917a
preventing the callbacks from opening the display.