qt: re-enable using top-level window for tool tip on Wayland with Qt 6.9.1
The behavior seems fine to me with Qt 6.9.0-beta3 and KWin Wayland 6.3.1.

Unfortunately there is a regression that the window x can no longer be
negative (QTBUG-135158) with regard to the interface window, but that
is not relevant to QTBUG-131899. This regression is expected to be fixed
by the latest Qt 6.9.1.