macosx: VLCTimeField: do not use identifier
Overriding setIdentifier is something that should not be done and
also that identifier is supposed to be unique so it is not really
suitable for our use-case here.

Additionally on older macOS versions this selector is sometimes called
with a nil identifier leading to a crash.

Just go back to a dedicated property for this without hooking into
the view identifier.