qml: downcast items from `QObject` to `QQuickItem` in `ExpandGridView.qml`
As their names suggest, they are supposed to be `QQuickItem` pointer.

Unfortunately when `Loader` started to gain ability to load `QObject`
in Qt 5:

> Since QtQuick 2.0, Loader can load any object type.

They have not changed the name of the property name `item` to `object`,
which can be confusing. I suppose they re-purposed `item` as a generic
"item", not necessarily `QQuickItem`. Still, this goes in conflict with
`Instantiator`'s `object` (not named `item`) (which always had the
ability to load `QObject`).