qml: cut too long empty indicator texts in media library views
It is planned to disable word wrap in `EmptyLabel`,
in an attempt to have sensible implicit size. This
allows us to use minimum width approach instead in
the main view.

There is little use for being responsive (via word
wrap) in the empty indicator. Usually views should
have minimum size which is at least the empty
indicator size (similar to the playlist pane). Word
wrap does not look good, the words can stack up and
down, making a column of many words.

Instead, I propose cutting the text, so that we
can have sensible minimum size (width). Currently
the window already has reasonable minimum width, and
there is also `isScreenSmall` which makes the playlist
pane be placed on top of the main view (which makes
the main view have as much width as the window has).