qml: create the scroll bar depending on the orientation in `ListViewExt`
By default we expect the viewport to cover the content width or height
depending on the orientation, so in normal cases we only need one scroll
bar. This behavior can obviously be overridden where the list view is
reused.

This fixes multiple things:
- Not being able to interact with the bottom part of the list view even
if the orientation is vertical and the viewport covers the whole content
width, because the scroll bar may still be technically visible and consume
events even though it does not display anything visually.
- Flashing scroll bar upon creation of the view until the content and
viewport sizes are determined (regarding the orientation where scroll bar
is not wanted).