qml: make scroll bar invisible if it does not show anything
This makes it possible to not consume the events when there
is nothing shown. We can do this because in this scroll bar
style when `size` is smaller than 1.0, the handle is always
visible. So this change is mostly relevant when `size` is
1.0, which means viewport size covers the content size.

This is useful for situations such as, playlist height is
equal or greater than the combined height of the items added
to it. With this patch, clicking on the items where the scroll
bar would be shown makes it possible for items to react to
the events (such as, selecting or showing the context menu).

I did not use `background.visible` or `contentItem.visible`
because they already depend on the control's visibility (as
if parent is invisible the children would be invisible).