contrib/src: use HTTPS for most contribs
This patch updates the URLs for contrib retrieval from
HTTP to HTTPS, where possible. URls that could not be
moved to HTTPS are marked with an XXX comment.

One domain is additionally changed: libsamplerate is
now retrieved from the libsndfile organization
on GitHub instead of mega-nerd.com, as the latter
does not offer HTTPS. The former was confirmed
for release integrity against the latter and was
cross-checked as the living source for libsamplerate
against Homebrew and Arch Linux, both of which use
theGitHub organization.

Signed-off-by: William Woodruff <william@trailofbits.com>