CI: update the Android docker image with NDK 27c
It can supports 16KB aligned page sizes with a flag in the NDK build package. [^1]

It is not on by default so we can test things without being stuck with this image.

The relevant change will need to be done on the libvlcjni side, including switching to this image as well. This is done in [^2].

[^1]: https://developer.android.com/guide/practices/page-sizes#compile-r27
[^2]: https://code.videolan.org/videolan/libvlcjni/-/merge_requests/112