vout: android: revert window to base values
This fixes OpenGL vout layout when it is used just after with the same
Android Surface.

This android vout is normally only used with opaque chroma (mediacodec),
and the opengl one should be used for software rendering.

Nevertheless, this vout could be used for software rendering when the
OpenGL one failed. This was the case when the avcodec AV1-hw decoder
requested a VAAPI chroma (via a YUV chroma fallback).

Fixes #29120