qt: introduce `VoronoiSnow.frag`
I initially wanted to use an already made snow shader,
as there are better ones. However, I could not find a
good one with compatible license, so I had to come up
with one myself. Thankfully Inigo Quilez's voronoi
and snowflake SDF helped me (MIT license).

This is not the best snow shader effect, and I can not
create a good one because I'm not familiar with GLSL
or graphics programming in general. But it seemed decent
to me for the use case here.

Note that this is additive, so if the background is
pure white the snow would not be visible.

I also investigated whether to use Qt's particle system
for this, instead of a fragment shader. I decided to
do this with a fragment shader because we currently do
not use the particle system, the performance benefit is
not clear, and the fragment shader is simpler (thanks to
Inigo Quilez).