qml: fix dropping item in playlist view may not open new playlist dialog
This is because `itemContainsDrag` may be undefined or null, which is
observed when the list is empty.

With the chaining operator, the evaluation yields undefined, and then
the function can handle the passed undefined index (which is already
the case here with `acceptDropFunc`, where
`DialogsProvider.playlistsDialog()` is called when the passed index is
undefined).