qt: make `PlaylistsDialog` no longer a singleton
Dialogs that carry the state of their initiators
should not be singleton. In this case, this dialog
always needs some sort of media to be functional.

This is already the case with the media info
dialog, for example, current media info dialog
is a singleton but media info dialog of a specific
media is not.

Since this dialog is always tied to some specific
media, I see no point of this being a singleton.
Its title suggests that this dialog is only meant
to be used for adding items to an existing or new
playlist.

Asynchronousity is also a problem when it is a
singleton, what happens if the user wants to add
items before the previous (asynchronous) addition
is not complete yet?