qt: when possible segregate playlists in playlist dialog
Currently the interface only provides "video" and "audio"
playlists. Mixed playlists don't seem to be presented.
This is something that can be handled trivially with drag
and drop case.

Adding to playlist action in context menu is more of a
problem, because in that case there is no specific open
playlist view that we can pick as the target.

I'm not sure what is the best behavior with regard to
the latter situation, but I assume we need to segregate
there too, because if the user can not make use of a
mixed playlist, there is no point creating them or
appending media.

I followed a naive approach, simply used `showPlayAsAudioAction`.
However, it could be possible to:
- If there is a single item, check if it is audio or video, and
  pick the appropriate type playlist.
- If there are multiple items, always pick `PLAYLIST_TYPE_ALL`.
- If there are multiple items, iterate through all items, and
  if all of them are the same, pick the appropriate playlist
  (i.e. video, or audio playlist), if there are different type
  of items, pick `PLAYLIST_TYPE_ALL`.