qml: fix broken customized round button
This customization has been broken for a while, I
have not previously tried to fix it because there
were plans of getting rid of it (at least, the
case with horizontal list view). However, not only
the buttons still remain there, but it is now also
used in more places (be87f3ea).

Customization is not worth it if the control is
never or barely used as in these cases we can simply
rely on Qt Quick Basic style directly, and adjust
the palette and sizes and be done with it.

Note that, we can expect Qt Quick Controls Basic
style to be available as long as Qt Quick Controls
itself is available (which we require). This means
that we should be able to freely import the
`QtQuick.Controls.Basic` QML module.