contrib: flac: Include -pthread in the pkg-config file when necessary
This fixes linking with llvm-mingw (or any mingw distribution that
doesn't default to linking in winpthreads).

This wasn't noticed in the upstream VLC CI, as it builds with
extras/packages/win32/build.sh, which passes in -Wl,-l:libpthread.a
in VLC_LDFLAGS (in order to prefer linking winpthreads statically,
rather than dynamically, for libraries that do use winpthreads).

This patch has been sent upstream in
https://github.com/xiph/flac/pull/832.