egl_gbm: fix dma-buf file descriptor leak
Gbm states the following:

> Each call to gbm_bo_get_fd_for_plane() returns a new
> file descriptor and the caller is responsible for
> closing the file descriptor.

This is currently not satisfied. Because of that, the
application quickly runs out of the file descriptors
that the system can give when resizing the video window.

This fixes the crash for me when I resize the window on
wayland. In my case it takes a few seconds of accumulated
resize before running out of file descriptors.