player: fix vlc_player_GetTime() while paused
In the case where an ES did not send an update after being paused
(likely when audio only), the last timer point was still valid causing
the Interpolate function to continue.

Do check for paused before interpolating a point.

Fixes #28353

The issue was only visible when using vlc_player_GetTime() and
libvlc_media_player_get_time() (so, no problem for Qt that is using the
timer API).