qt: use `QWidget::show()` instead of `QDialog::open()` for playlists dialog
So that the dialog respects its modality. The add to playlist dialog should
be modeless, especially now that it is no longer a singleton.

I previously used `QDialog::open()` because it is specific to `QDialog`,
unlike `QWidget::show()`, but I did not realize that `QDialog::open()` does
not respect the modality and always shows the dialog modal.