demux: avi: cut QNap title on first invalid character
We should only use valid UTF-8 metadata.
We don't know if this is supposed to be ASCII, UTF-8 or something else.
This is the same code as EnsureUTF8() except we cut the rest of the string if it's invalid.

Ref. #28976

(cherry picked from commit 9c5ea58677b9f86e278a4c2c0b0fa738bf72ccd4)