qt: use move assignment to simplify construction in `ColorizedSvgIcon`
This also removes the overzealous assertion of a valid SVG icon engine
plugin being retrieved, which may not be the case if relevant package
(Qt Svg) is not installed.

Since we are using the assignment operator in both branches, I don't
think it should be necessary to call the `QIcon` constructor anymore.

Unlike the image colorization, where a purple image is provided if the
engine is not available, I decided to rely on the default behavior of
`QIcon` in this case, which I assume is having no icon.