src: meson: add support for _vlc_static_module on Darwin
The static module bank, represented by the _vlc_static_module symbol, is
conditionally enabled depending on the link options in configure.ac.
Indeed, in the past with bitcode support and maybe today with LTO
support, there is no way to use the dynamic_lookup option from the
linker, and the two-level namespace from darwin blocks the symbol from
being looked up in the executable.

_vlc_static_module is in particular used by the tests, so re-expose the
options to enable it by default. The define was missing in particular.