splitter: add shortcut for "splitter"
The module is used by name in the codebase, so ensure the name is
actually available for the module.

Both Meson and Autoconf compile the module as "video_splitter_plugin",
but Autoconf doesn't use CPPFLAGS, which makes the module being named
"splitter" since it only compiles "splitter.c". But Meson will correctly
name the module as "video_splitter".