qml: handle item reusing properly in `MediaCover`
The documentation advises to free resources when the item is
pooled. In this case, we do not know if the pooled item is
going to be reused with the same image, so I release the
image. I do not set the texts to empty, as any potential
resource consumption there is considered negligible, at
least for now.

The documentation also mandates to reset the delegate state
when the item is reused. In this case, there is not much to
reset because of declarative bindings. However, there is
`_loadTimeout` that needs to be reset, in addition to the
restoration of the broken bindings used to free the resources.