qt: assume the caller thread does not change in `VideoSurface::synchronize()`
We can have this assumption to not check for the thread each time it is
called. This can be expensive, as the method is called at v-sync rate.

We can do this because `VideoSurfaceProvider::supportsThreadedSurfaceUpdates()`
is already probed once and assumed to be constant.

For safety, I have added assertions for the debug build to check if this
assumption does not hold.