codec/cc: Fix garbled DVD CEA-608 captions with large cc_counts
The GOP parser for captions when found in DVD GOPs had an error
when determining the CC count, which excluded the top bit (i.e.
only using the lower four bits of a five bit field).  Thus
it would work properly if the cc_count was less than 16, but would
lose the rest of the captions in the GOP.

Parse all five bits of the field.

Signed-off-by: Devin Heitmueller <dheitmueller@ltnglobal.com>