codec/cc: Fix lost CEA-608 pair for DVD captions with truncate flag
If parsing GOP based captions, the truncate field wasn't being
properly respected, resulting in the last pair being lost if the
flag was set.  This would cause intermittent caption bytes to be
missing (e.g. missing characters in rendered captions).

Correct the logic so that the last pair isn't lost, by comparing
the code against the original ccextractor code it was based on.

Tested with "Sports Night", season 1, disc 1.

Signed-off-by: Devin Heitmueller <dheitmueller@ltnglobal.com>