qml: automatically request focus in `MainViewLoader` loaded item
So that we don't need to use `focus: true` in every main view.
We can do this because main views are expected to request focus.

Currently views in `VideoAll` are missing this, instead of setting
them there, I thought that this approach would be less fragile.

Focus reason is a different matter and is not handled here.