qml: remove the built-in padding from `TableViewDelegateExt`'s content item
This causes alignment issues with the header. It seems that they were added
in dec5d4b6, as part of avoiding directly referencing content item, but it
is not clear to me why at the moment and I do not remember why it was added
back then. It is likely a remnant from the idea of having a default padding
that act as the base padding and the root control's padding acting as extra
padding. There is still not a good solution for this in QML, i.e. adjusting
property when an item is reused while keeping the default value in account.
However, we have nevertheless moved away from that approach for a while, as
it is better to have full control of something under the jurisdiction of an
item where it is reused. If necessary, the default value can be repeated in
where it is reused.

That being said, I don't remember having this issue before, so it is likely
some change in the header caused this alignment issue. Perhaps the header
also contained this same amount of base padding, and it was removed there
but kept here.