qml: delay the visibility binding inside layout in `ArtworkInfoWidget`
This fixes a potential polish loop. I reproduce this when
I change the user scale of the interface, especially in a
rapid succession.

Having the binding delayed here makes more sense than
not having it, because it depends on something that it
also has an effect. And we already delay the visibility
bindings in layouts in other cases.