qml: consider display margins properly in hover area placement of `FadingEdgeForListView`
The display "margin" grows outwards the viewport, this can also be seen in `FadingEdge`
where its `ShaderEffectSource` use the negative of the display margin for its position
(left and top margins). The hover areas must perfectly overlap with the fading areas,
we need this change to satisfy that.

I also fix mistakenly using beginning margin for the end hover area.