display: expose vout_display_SetFormat API
The vout display API currently store an internal state in the core,
tracking what was previously requested to the vout_display by previous
clients.

Recently, this state has been used to provide a vout_display place to
the vout_display modules which used it to place the video frame and the
subpictures on screen. This enforced the usage of a core vout_display
API to manipulate the vout_display objects.

However, there were no vout_display core API to update the format of the
displays, and calling the callback function directly led to
discrepancies between the internal core state and the format visible to
the vout display through the pictures.

This commit introduces a core API to circumvent that, but later commits
will eventually get rid of the internal state, like it's done for the
vout_window implementation.