qt: map `PLAYLIST_TYPE_AUDIO` to `VLC_ML_PLAYLIST_TYPE_AUDIO`
With the `ONLY` option, playlists can disappear in the interface
when the user adds a different type content, such as by dragging.
This is problematic because the interface may not know whether
the dragged media is considered as video or audio by media library,
so it is not possible to (even if we want) disallow (at least
from the interface side) adding different type of media. Not to
mention, media library can be manipulated by anything, so such
blockage would probably does not make sense.

This is especially important because we currently do not offer a
view that shows all or mixed playlists.