qt: adapt exit code after libvlc_Quit() fix on Darwin
The Abort() function was set as a handler for the libvlc_Quit()
callback, through libvlc_SetExitHandler, and was used to provide Qt
runloop termination when the application was exiting.

But terminating the runloop and Qt resources was not correct since it
didn't account for the internal resource reference counting. Instead,
now that darwinvlc.m correctly handles libvlc termination by releasing
the interfaces before killing the runloop, we can just use the normal
code path for interfaces, which in the case of Qt already handles the
reference counting, and wait for the Qt runloop to exit.

Fix #28917